Update index.js
When he noticed how the guests picked the places of honor at the table, he told them this parable: “When someone invites you to a wedding feast, do not take the place of honor, for a person more distinguished than you may have been invited. If so, the host who invited both of you will come and say to you, ‘Give this person your seat.’ Then, humiliated, you will have to take the least important place. But when you are invited, take the lowest place, so that when your host comes, he will say to you, ‘Friend, move up to a better place.’ Then you will be honored in the presence of all the other guests. For all those who exalt themselves will be humbled, and those who humble themselves will be exalted.”